was able to get my models migrated
