+++ FamilyCB/FamilyCBProject/models/__init__.py
@@ -1,0 +1,6 @@
+from .comment import Comment
+from .family import Family
+from .ingredient import Ingredient
+from .member import Member
+from .recipe import Recipe
+from .modelfactory import model_factory
