+++ FamilyCB/FamilyCBProject/views/members/details.py
